+++ tests/platform_helper/domain/test_conduit.py
@@ -8,9 +8,9 @@ from dbt_platform_helper.exceptions import AddonNotFoundError
 from dbt_platform_helper.exceptions import CreateTaskTimeoutError
 from dbt_platform_helper.exceptions import ECSAgentNotRunning
 from dbt_platform_helper.exceptions import InvalidAddonTypeError
-from dbt_platform_helper.exceptions import NoClusterError
 from dbt_platform_helper.exceptions import ParameterNotFoundError
 from dbt_platform_helper.exceptions import SecretNotFoundError
+from dbt_platform_helper.providers.ecs import NoClusterError
 from dbt_platform_helper.utils.application import Application
 from dbt_platform_helper.utils.application import Environment
