+++ tests/platform_helper/providers/test_ecs.py
@@ -5,8 +5,8 @@ import pytest
 from moto import mock_aws
 
 from dbt_platform_helper.exceptions import ECSAgentNotRunning
-from dbt_platform_helper.exceptions import NoClusterError
 from dbt_platform_helper.providers.ecs import ECS
+from dbt_platform_helper.providers.ecs import NoClusterError
 from tests.platform_helper.conftest import mock_parameter_name
 from tests.platform_helper.conftest import mock_task_name
