+++ tests/platform_helper/test_exceptions.py
@@ -12,11 +12,11 @@ from dbt_platform_helper.exceptions import CreateTaskTimeoutError
 from dbt_platform_helper.exceptions import ECSAgentNotRunning
 from dbt_platform_helper.exceptions import ImageNotFoundError
 from dbt_platform_helper.exceptions import InvalidAddonTypeError
-from dbt_platform_helper.exceptions import NoClusterError
 from dbt_platform_helper.exceptions import NoCopilotCodebasesFoundError
 from dbt_platform_helper.exceptions import NotInCodeBaseRepositoryError
 from dbt_platform_helper.exceptions import ParameterNotFoundError
 from dbt_platform_helper.exceptions import SecretNotFoundError
+from dbt_platform_helper.providers.ecs import NoClusterError
 
 
 @pytest.mark.parametrize(
